requirements.txt
requests>=2.31,<3
pandas>=2.1,<3
numpy>=1.25,<3
anthropic>=0.40,<1
python-dotenv>=1.0,<2
websocket-client>=1.6,<2
